requirements.txt
hdbcli
pandas
python-dotenv